class_project/MSML610/Fall2025/Projects/UmdTask40_Fall2025_Whisper-Large-V3_Multilingual_Customer_Support_Chatbot/Term2025/projects/TutorTask40_Fall2025_Whisper_Large_V3_Multilingual_Customer_Support_Chatbot/requirements.txt
transformers>=4.40.0
accelerate
sentencepiece
openai-whisper
soundfile
protobuf==3.20.*
huggingface-hub
numpy
torchaudio
streamlit
